+++ src/redux/configureStore.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 import BookSlice from './books/books';
-import CategorySlice from './categories/categories';
+import CategorySlice from './category/category';
 
 export default configureStore({
   reducer: {
